Move poorly named modules to better locations
"project" is overloaded with Project objects and is potentially changing
name in the platform anyway.

"pipeline" could describe a research pipeline, auth pipeline, or several
other things tangentially related to this project.
