+++ jobserver/views/job_requests.py
@@ -15,7 +15,7 @@ from ..backends import backends_to_choices
 from ..forms import JobRequestCreateForm, JobRequestSearchForm
 from ..github import get_branch_sha
 from ..models import Backend, JobRequest, User, Workspace
-from ..project import get_actions, get_project, load_yaml, render_definition
+from ..pipeline_config import get_actions, get_project, load_yaml, render_definition
 from ..utils import raise_if_not_int
